pyproject.toml
[project]
name = "ont-2025"
version = "0.2.0"
description = "YouTube 视频提取 PPT：模块化流水线 + GUI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "yt-dlp>=2024.1.0",
    "extract-video-ppt>=1.1.0",
    "fpdf2>=2.0,<2.8",
    "pymupdf>=1.27.1",
    "PySide6>=6.6",
]
